refactor: split EXPLAIN renderer into explain.rs; factor indexed_equality_candidates
cypher/mod.rs crossed its 500-line module cap after the ClosureProbe rows;
the static plan renderer moves out whole. filter_nodes_by_conditions gets its
equality fast path extracted. Behavior identical (same suite counts).
Source-quality exceptions tightened, api baselines gain label_cardinality.
